[Refactor] Replaced masked Java exceptions for specific OpenMarkov exception. [Refactor] Removed manually setting index of enum constants, as it led to throwing exceptions on unreachable code (Said unreachable code now doesn't need to exist, eliminating said unreachable exceptions). [Test] Added a tag to prevent certain tests from being run in parallel with OpenMarkov's run configurations. [Exception, Test] Most of the exceptions thrown by calling Constructor#newInstance(Object...) are no longer reachable, as they are verified by the test InheritanceFiltersAreMet. [Bugfix] Partially fixed bug from issue 533 (Cloning issues in class Node). [Refactor] Code quality to reduce exceptions.

+++ src/main/java/org/openmarkov/gui/action/ICITablePotentialValueEdit.java
@@ -218,7 +218,7 @@ import java.util.List;
 		if (!leakyFlag) {//noisy parameters
 			newNoisyParameters[position] = Util.roundAndReduce(newValue, epsilon, maxDecimals);
 			while (listIterator.hasNext() == true) {
-				priorityListPosition = (Integer) listIterator.next();
+                priorityListPosition = listIterator.next();
 				sum = Util.roundAndReduce(sum + newNoisyParameters[priorityListPosition], epsilon, maxDecimals);
 				//sum = sum + newNoisyParameters[priorityListPosition];
 				//sum = roundingDouble(sum + newNoisyParameters[priorityListPosition]);
@@ -231,7 +231,7 @@ import java.util.List;
 			if (sum > 1.0) {
 				listIterator = priorityList.listIterator();
 				while (listIterator.hasNext() == true && rest != 0) {
-					priorityListPosition = (Integer) listIterator.next();
+                    priorityListPosition = listIterator.next();
 					//rest = rest - newNoisyParameters[priorityListPosition];
 					rest = Util.roundAndReduce(rest - newNoisyParameters[priorityListPosition], epsilon, maxDecimals);
 					//rest = rest - newTable[pos];
@@ -255,7 +255,7 @@ import java.util.List;
 		} else if (leakyFlag) {//leaky parameters
 			newLeakyParameters[position] = Util.roundAndReduce(newValue, epsilon, maxDecimals);
 			while (listIterator.hasNext() == true) {
-				priorityListPosition = (Integer) listIterator.next();
+                priorityListPosition = listIterator.next();
 				sum = Util.roundAndReduce(sum + newLeakyParameters[priorityListPosition], epsilon, maxDecimals);
 				//sum = roundingDouble(sum + newLeakyParameters[priorityListPosition]);
 				//sum += newTable[pos];
@@ -267,7 +267,7 @@ import java.util.List;
 			if (sum > 1.0) {
 				listIterator = priorityList.listIterator();
 				while (listIterator.hasNext() == true && rest != 0) {
-					priorityListPosition = (Integer) listIterator.next();
+                    priorityListPosition = listIterator.next();
 					rest = Util.roundAndReduce(rest - newLeakyParameters[priorityListPosition], epsilon, maxDecimals);
 					//rest = roundingDouble(rest - newLeakyParameters[priorityListPosition]);
 					//rest = rest - newTable[pos];
